v3.0.0-alpha.3
This adds *a lot*.

Finishes up the implementation of /config. The only thing missing is the language setting as I have not added that to the rest of the bot yet (and honestly i'm on the fence about doing that at all).

On the topic of /config, I couldn't find a way to update ephemeral messages so I just gave up and started to respond with an interaction follow-up instead. This isn't as neat as I'd like, but oh well. 🤷‍♀️

Next, it adds the info command, disables (until I get around to implementing it) /level and /leaderboard.

It also completely removes all music bot functionality, I do not have interest in maintaining that feature-set anymore.

Pushes all file-handling code to Robin, I honestly thought I already did this, but here we are.

I started an update to the 8ball command but decidedly gave up on that because as it turns out, making a generic response to non-yes/no questions is pretty hard. Even making answers that are comedically wrong is rather difficult.

I also finally fixed the bloody log4j2.xml file to make it less awful to debug issues.

The commands(.yml) file has been updated to feature the new overview field from Robin v1.1. You will need to regenerate or manually update your config file.

The config(.yml) has a fully fleshed out series of options for ClaireData, so be sure to add that section or regenerate your config file.

Besides some other misc. bugfixes, I think that's about it.

P.S. the longest part of all of this was figuring out what I changed 3 months ago and then left partially complete. Probably don't do that.