Alpha 5
The user defaults for the database have been updated. It is *highly* recommended that you update your config file to follow or you will face a swathe of errors.

This commit mostly implements the level/leaderboard system. However, it is wildly untested beyond checking that it won't absolutely nuke the production database.