Initial work for ClaireLang
Finally getting around to implementing this. If I don't implement it before my planned feature set for v3.5, it will never happen as the complexity of implementing this will go through the roof in the update as planned.

Putting this in a separate branch so that when I have downtime at work, I can work on the slow and painful process of migrating everything to this system.