Probably a bad idea
I prefer the idea of this structure, I just don't like the implementation of this structure.

Also considered doing one giga-class (technical term :)), but that would've been much less readable than whatever in the fuck this is.

Probably would've been easier to understand the intent of though... I will need to write some really good javadoc if I stick with this.