feat: command embeds fully localized
Almost entirely courtesy of GPT-5. This wouldn't be finished anytime soon without LLM assistance. The hard part was done by hand, the machine can do the chores.

Now the only remaining work (that immediately comes to mind) is the MessageComponents embeds. It looks like I still need to add most (or all) of those fields to the language file so that's gonna wait.